tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/080.html

<!DOCTYPE html>
<html><head>
	<title>scheduler: IFRAMEs added with DOM (appendChild), javascript: URL</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>
	<div></div>
	<script>log('inline script #1');
		var iframe=document.createElement('iframe');
		iframe.src='javascript:parent.log(\'JS URL\');\'<html><script>parent.log(\\\'frame script\\\')<\/script></html>\'';
		document.getElementsByTagName('div')[1].appendChild(iframe);
		log('end script #1');
	</script>

	<script type="text/javascript">
	log( 'inline script #2' );
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['inline script #1', 'end script #1', 'inline script #2', 'JS URL', 'frame script']);
		t.done();
	}
	onload = t.step_func(function(){setTimeout(t.step_func(test), 400);})
	</script>

</body></html>
